Fix retaggings from union and retag all cloned vars
